requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-polymorphic==3.1.0
django-render-partial==0.4
pillow==11.0.0
python-decouple==3.8
setuptools==75.2.0
sqlparse==0.5.1
